Rename to pascal-cased name
