+++ src/components/cluster/detail/AddNodePool.js
@@ -4,7 +4,7 @@ import { hasAppropriateLength } from 'lib/helpers';
 import { RadioWrapper } from '../new/CreateNodePoolsCluster';
 import AvailabilityZonesParser from './AvailabilityZonesParser';
 import AWSInstanceTypeSelector from '../new/AWSInstanceTypeSelector';
-import NodeCountSelector from 'shared/node_count_selector';
+import NodeCountSelector from 'shared/NodeCountSelector';
 import produce from 'immer';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
