+++ src/components/cluster/detail/ScaleClusterModal.js
@@ -7,7 +7,7 @@ import BootstrapModal from 'react-bootstrap/lib/Modal';
 import Button from 'UI/button';
 import ClusterIDLabel from 'UI/cluster_id_label';
 import cmp from 'semver-compare';
-import NodeCountSelector from 'shared/node_count_selector';
+import NodeCountSelector from 'shared/NodeCountSelector';
 import PropTypes from 'prop-types';
 import React from 'react';
