+++ src/components/cluster/detail/ScaleNodePoolModal.js
@@ -6,7 +6,7 @@ import { FlashMessage, messageTTL, messageType } from 'lib/flash_message';
 import BootstrapModal from 'react-bootstrap/lib/Modal';
 import Button from 'UI/button';
 import ClusterIDLabel from 'UI/cluster_id_label';
-import NodeCountSelector from 'shared/node_count_selector';
+import NodeCountSelector from 'shared/NodeCountSelector';
 import PropTypes from 'prop-types';
 import React from 'react';
